+++ plugins/module_utils/_elbv2/listeners.py
@@ -26,8 +26,33 @@ if typing.TYPE_CHECKING:
 
 from . import api as _api
 from . import rules as _rules
+from .common import AnsibleELBv2Error
 from ..elb_utils import convert_tg_name_to_arn
 from ..transformation import scrub_none_parameters
+
+
+def validate_listener_https_requirements(listeners: Optional[List[Dict[str, Any]]]) -> None:
+    """
+    Validate that HTTPS listeners have required SSL configuration.
+
+    Checks that listeners using the HTTPS protocol include both SslPolicy
+    and Certificates configuration.
+
+    Args:
+        listeners: List of listener configuration dicts, or None
+
+    Raises:
+        AnsibleELBv2Error: If any HTTPS listener is missing required SslPolicy or Certificates
+    """
+    if listeners is None:
+        return
+
+    for listener in listeners:
+        if listener.get("Protocol") == "HTTPS":
+            if listener.get("SslPolicy") is None:
+                raise AnsibleELBv2Error(message="'SslPolicy' is a required listener dict key when Protocol = HTTPS")
+            if listener.get("Certificates") is None:
+                raise AnsibleELBv2Error(message="'Certificates' is a required listener dict key when Protocol = HTTPS")
 
 
 def _compare_listener(current_listener: Dict[str, Any], new_listener: Dict[str, Any]) -> Optional[Dict[str, Any]]:
